+++ tests/mock_server.py
@@ -82,7 +82,8 @@ def create_app():
             return json.dumps({
                 "data": {
                     "viewer": {
-                        "entity": "vanpelt"
+                        "entity": "vanpelt",
+                        "flags": '{"code_saving_enabled": true}'
                     }
                 }
             })
@@ -158,7 +159,7 @@ def create_app():
 
     @app.errorhandler(404)
     def page_not_found(e):
-        print("Got request to: %s" % e)
+        print("Got request to: %s" % request.url)
         return "Not Found", 404
 
     return app