+++ wandb/apis/internal.py
@@ -324,6 +324,7 @@ class Api(object):
         query Viewer{
             viewer {
                 id
+                flags
                 entity
                 teams {
                     edges {