+++ wandb/run_manager.py
@@ -279,7 +279,8 @@ class FileEventHandlerConfig(FileEventHandler):
 
     def finish(self):
         if self._thread:
-            self._thread.join()
+            # Cancel the current thread to keep moving
+            self._thread.cancel()
             self._thread = None
 
         self._update()
@@ -665,11 +666,9 @@ class RunManager(object):
         """Stops file syncing/streaming but doesn't actually wait for everything to
         finish. We print progress info later.
         """
-
         # TODO: there was a case where _file_event_handlers was getting modified in the loop.
         for handler in list(self._file_event_handlers.values()):
             handler.finish()
-
         self._file_pusher.finish()
         self._api.get_file_stream_api().finish(exitcode)
         # In Jupyter notebooks, wandb.init can be called multiple times in the same
@@ -1047,7 +1046,7 @@ class RunManager(object):
 
         env = self._run.set_environment(environment=env)
 
-        if not env.get(wandb_env.DISABLE_CODE):
+        if wandb_env.should_save_code():
             logger.info("saving patches")
             self._api.save_patches(self._run.dir)
         if env.get("SPELL_RUN_URL"):
@@ -1079,11 +1078,12 @@ class RunManager(object):
         if self._run_status_checker:
             self._run_status_checker.shutdown()
 
+        self._run.history.close()
+
         if self._cloud:
             logger.info("stopping streaming files and file change observer")
             self._end_file_syncing(exitcode)
 
-        self._run.history.close()
 
     def run_user_process(self, program, args, env):
         """Launch a user process, capture its output, and sync its files to the backend.